requirements.txt
pysnow
python-dotenv
requests